Add authentication in new file
