+++ app.js
@@ -9,7 +9,6 @@ var bodyParser = require('body-parser');
 const session = require('express-session')
 const passport = require('passport')
 const RedisStore = require('connect-redis')(session)
-const GoogleStrategy = require('passport-google-oauth20').Strategy
 
 var index = require('./routes/index');
 var users = require('./routes/users');
@@ -19,96 +18,7 @@ const config = require('./config')
 
 var app = express();
 
-
-app.use(bodyParser.urlencoded({
-  extended: false
-}))
-
-
-passport.authenticationMiddleware = function () {
-  return function (req, res, next) {
-    if (req.isAuthenticated()) {
-      return next()
-    }
-    res.redirect('/')
-  }
-}
-
-app.use(passport.initialize())
-
-// extract google profile
-function extractProfile (profile) {
-  var imageUrl = '';
-  if (profile.photos && profile.photos.length) {
-    imageUrl = profile.photos[0].value;
-  }
-  var ret = {
-    id: profile.id,
-    displayName: profile.displayName,
-    image: imageUrl
-  };
-  return ret;
-}
-
-passport.serializeUser((user, cb) => {
-  cb(null, user);
-});
-
-passport.deserializeUser((obj, cb) => {
-  cb(null, obj);
-});
-
-// Use the GoogleStrategy within Passport.
-//   Strategies in passport require a `verify` function, which accept
-//   credentials (in this case, a token, tokenSecret, and Google profile), and
-//   invoke a callback with a user object.
-passport.use(new GoogleStrategy({
-    clientID: config.google.clientID,
-    clientSecret: config.google.clientSecret,
-    callbackURL: config.google.callbackURL,
-    accessType: 'offline',
-  }, (accessToken, refreshToken, profile, cb) => {
-  // Extract the minimal profile information we need from the profile object
-  // provided by Google
-    cb(null, extractProfile(profile));
-  }
-));
-
-app.get(
-  // Login url
-  '/auth/google',
-
-  // Save the url of the user's current page so the app can redirect back to
-  // it after authorization
-  (req, res, next) => {
-    if (req.query.return) {
-      req.session.oauth2return = req.query.return;
-    }
-    next();
-  },
-
-  // Start OAuth 2 flow using Passport.js
-  passport.authenticate('google', { scope: ['email', 'profile'] })
-);
-// [END authorize]
-
-// [START callback]
-app.get(
-  // OAuth 2 callback url. Use this url to configure your OAuth client in the
-  // Google Developers console
-  '/auth/google/callback',
-
-  // Finish OAuth 2 flow using Passport.js
-  passport.authenticate('google'),
-
-  // Redirect back to the original page, if any
-  (req, res) => {
-    const redirect = req.session.oauth2return || '/';
-    delete req.session.oauth2return;
-    res.redirect(redirect);
-  }
-);
-
+// session stuff
 app.use(session({
   store: new RedisStore({
     url: config.redisStore.url
@@ -118,13 +28,28 @@ app.use(session({
   saveUninitialized: false
 }))
 
+app.use(passport.initialize())
 app.use(passport.session())
+// end session stuff
 
-app.get('/logout', function(req, res){
-  req.logout();
-  res.redirect('/');
+// authentication stuff
+app.use(require('./routes/auth.js').router)
+
+// render the index page and any of the static files without authentication
+app.use('/', index);
+app.use(express.static(path.join(__dirname, 'public')));
+
+// only let through the requests that are authenticated after this.
+app.use(function (req, res, next) {
+  if (req.isAuthenticated()) {
+    return next()
+  }
+  res.redirect('/')
 });
 
+app.use(bodyParser.urlencoded({
+  extended: false
+}))
 
 // Setup websocket connection
 var server = require('http').Server(app);
@@ -165,9 +90,7 @@ app.use(logger('dev'));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(cookieParser());
-app.use(express.static(path.join(__dirname, 'public')));
 
-app.use('/', index);
 app.use('/users', users);
 app.use('/ws', ws);
