+++ routes/index.js
@@ -2,8 +2,9 @@ var express = require('express');
 var router = express.Router();
 
 /* GET home page. */
-router.get('/', function(req, res, next) {
-  res.render('index', { title: 'Express' });
+router.get('/dashboard', function(req, res, next) {
+  console.log(req.user);
+  res.render('dashboard', { user: req.user });
 });
 
 module.exports = router;
